encapsulate code to get mints info
